Refactor: move `ScoreDto` to `dto` package
